refactor(exports): extract service layer from route (WP1.8)
Move mapping tables, dataframe transforms, query construction, sheet assembly,
and export-to-log persistence out of routes/exports.py into a new
utils/export_service.py. The route handlers stay thin (parse/validate →
delegate → same responses, including file/download responses).

Behavior-preserving:
- OD3: the export-to-log exercise_order side effect (recalculate_exercise_order)
  is preserved exactly; endpoint stays GET (the GET→POST fix is WPB.3, later).
- OD10: streaming-threshold helpers + workbook mechanics remain in
  utils/export_utils.py and are homed here (MAX_EXPORT_ROWS, stream_export_rows,
  stream_excel_response); nothing deleted.
- No timing changes (A7's sleep removal state preserved).
- Service uses DatabaseHandler + get_logger() and does NOT import from routes
  (a local _column_exists copy avoids the routes dependency; WP2.6 moves it to
  utils).

Gate: import smoke OK (routes.exports + utils.export_service). Full pytest +
plan export/download + workout-log import E2E run on CI.
